+++ crates/client/gateway/server/src/helpers.rs
@@ -2,10 +2,9 @@ use std::collections::HashMap;
 
 use hyper::{body::Incoming, header, Request, Response, StatusCode};
 use mp_block::{BlockId, BlockTag};
+use mp_gateway::error::{StarknetError, StarknetErrorCode};
 use serde::Serialize;
 use starknet_types_core::felt::Felt;
-
-use crate::error::{StarknetError, StarknetErrorCode};
 
 pub(crate) fn service_unavailable_response(service_name: &str) -> Response<String> {
     Response::builder()
@@ -116,10 +115,4 @@ pub(crate) fn block_id_from_params(params: &HashMap<String, String>) -> Result<B
 
 pub(crate) fn include_block_params(params: &HashMap<String, String>) -> bool {
     params.get("includeBlock").map_or(false, |v| v == "true")
-}
-
-impl From<StarknetError> for Response<String> {
-    fn from(error: StarknetError) -> Self {
-        create_json_response(hyper::StatusCode::BAD_REQUEST, &error)
-    }
 }