+++ crates/client/sync/src/fetch/mod.rs
@@ -4,10 +4,8 @@ use std::time::Duration;
 use futures::prelude::*;
 use mc_block_import::UnverifiedFullBlock;
 use mc_db::MadaraBackend;
-use mc_gateway::{
-    client::builder::GatewayProvider,
-    error::{SequencerError, StarknetError, StarknetErrorCode},
-};
+use mc_gateway_provider::GatewayProvider;
+use mp_gateway::error::{SequencerError, StarknetError, StarknetErrorCode};
 use mp_utils::{channel_wait_or_graceful_shutdown, wait_or_graceful_shutdown};
 use tokio::sync::{mpsc, oneshot};