+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- feat(fgw): added `add_transaction` for gateway client
 - fix(fgw): include `l1_to_l2_consumed_message` in L1 handler receipt
 - build: up starknet-rs, starknet-types, blockifier(v0.8.0), cairo
 - feat(rpc): added `getCompiledCasm` method